fix: update generate_chunk to support chunks of length >= 26
Previously, I was using the `choose_multiple` implementation from
IteratorRandom. This had undesirable behavior when the chunk length was
exactly 26, in that the chunk would end up being the alphabet, in order,
every time.

Switching the `choose_multiple` from SliceRandom kept the slice randomly
shuffled even at a chunk length of 26, but it didn't allow for chunks
longer than 26 characters because `choose_multiple` does not allow for
repetition.

The solution I've settled on for now is to generate the indices of
selected characters, and then convert them into the actual characters in
a 2nd pass, instead of trying to sample from the slice directly.